+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'gary'
+name             'cloud-dev'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures gary'
+description      'Installs/Configures cloud-dev for all your cloud dev needs'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'