+++ spec/default_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe 'gary::default' do
+describe 'cloud-dev::default' do
   let(:chef_run) do 
     run = ChefSpec::ChefRunner.new(platform: 'ubuntu', version: '12.04')
-    run.converge('gary::Default')
+    run.converge('cloud-dev::Default')
   end
   it 'installs dependecies' do
     expect(chef_run).to install_package('curl')